+++ services/users/manage.py
@@ -6,10 +6,12 @@ import unittest
 
 from flask.cli import FlaskGroup
 
-from project import app, db
+from project import create_app, db
+from project.api.models import User
 
 
-cli = FlaskGroup(app)
+app = create_app()
+cli = FlaskGroup(create_app=create_app)
 
 
 @cli.command('recreate_db')
