+++ services/users/project/tests/base.py
@@ -3,7 +3,10 @@
 
 from flask_testing import TestCase
 
-from project import app, db
+from project import create_app, db
+
+
+app = create_app()
 
 
 class BaseTestCase(TestCase):
