+++ services/users/project/tests/test_config.py
@@ -7,7 +7,10 @@ import unittest
 from flask import current_app
 from flask_testing import TestCase
 
-from project import app
+from project import create_app
+
+
+app = create_app()
 
 
 class TestDevelopmentConfig(TestCase):
